Start service move.
Comment out code to get compile to check expanded output.
