+++ deepwell/Cargo.toml
@@ -28,6 +28,7 @@ serde = { version = "1", features = ["derive"] }
 serde_json = "1"
 str-macro = "1"
 sqlx = "0.5"
+thiserror = "1"
 tide = "0.16"
 wikidot-normalize = "0.9"
