+++ deepwell/src/methods/mod.rs
@@ -20,14 +20,13 @@
 
 mod prelude {
     pub use crate::api::{ApiRequest, ApiResponse};
-    pub use crate::models::*;
     pub use crate::services::RequestFetchService;
     pub use crate::web::{utils::error_response, HttpUnwrap, ItemReference};
     pub use chrono::prelude::*;
     pub use sea_orm::entity::prelude::*;
     pub use sea_orm::{Condition, Set};
     pub use std::convert::TryFrom;
-    pub use tide::{Body, Error, Request, StatusCode};
+    pub use tide::{Body, Error, Request, Response, StatusCode};
 }
 
 pub mod locales;
