+++ deepwell/src/services/page/mod.rs
@@ -18,6 +18,7 @@
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
 
+use super::prelude::*;
 use crate::api::ApiServerState;
 
 #[derive(Debug)]
